+++ Taskfile.tests.yml
@@ -101,13 +101,6 @@ tasks:
         {{ .mocha_bin }} 'test/e2e/commands/node-pem-kill.test.ts'
         {{ .reporter_options_prefix }}-e2e-node-pem-kill.xml"
 
-  test-e2e-node-local-hedera:
-    cmds:
-      - "{{ .test_prefix }}=\"Mocha E2E Node Local Hedera Tests\"
-        {{ .reporter_prefix }}='coverage/e2e-node-local-hedera'
-        {{ .mocha_bin }} 'test/e2e/commands/node-local-hedera.test.ts'
-        {{ .reporter_options_prefix }}-e2e-node-local-hedera.xml"
-
   test-e2e-node-local-ptt:
     cmds:
       - "{{ .test_prefix }}=\"Mocha E2E Node Local PTT Tests\"
